+++ sdk/identity/azure-identity/inc/azure/identity/managed_identity_credential.hpp
@@ -43,6 +43,91 @@ namespace Azure { namespace Identity {
   };
 
   /**
+   * @brief The type of managed identity identifier depending on how the managed identity is
+   * configured.
+   *
+   * @remark This can either be system-assigned, or user-assigned which corresponds to an identifier
+   * that represents either client ID, resource ID, or object ID, depending on how the managed
+   * identity is configured.
+   */
+  enum class ManagedIdentityIdType
+  {
+    SystemAssigned,
+    ClientId,
+    ObjectId,
+    ResourceId,
+  };
+
+  /**
+   * @brief The type of managed identity and its corresponding identifier.
+   *
+   * @remark This class holds the type and unique identifier for either a system or user-assigned
+   * managed identity.
+   */
+  class ManagedIdentityType final {
+  private:
+    ManagedIdentityIdType m_idType;
+    std::string m_id;
+
+  public:
+    /**
+     * @brief Constructs the type of managed identity.
+     *
+     * @remark This defaults to ManagedIdentityIdType::SystemAssigned.
+     */
+    explicit ManagedIdentityType() : m_idType(ManagedIdentityIdType::SystemAssigned) {}
+
+    /**
+     * @brief Constructs the type of managed identity.
+     *
+     * @param idType The type of the managed identity identifier.
+     * @param id The value of the managed identity identifier. This can be either a client ID,
+     * resource ID, or object ID.
+     *
+     * @remark For ManagedIdentityIdType::SystemAssigned, the id must be an empty string.
+     *
+     * @remark Make sure the type of ID matches the value of the ID. For example, the client
+     * ID and object ID are NOT interchangeable, even though they are both Uuid values.
+     */
+    explicit ManagedIdentityType(ManagedIdentityIdType idType, std::string id)
+        : m_idType(idType), m_id(id)
+    {
+      if (idType == ManagedIdentityIdType::SystemAssigned && !id.empty())
+      {
+        throw std::invalid_argument(
+            "There is no need to provide an ID (such as client, object, or resource ID) if you are "
+            "using system-assigned managed identity.");
+      }
+    }
+
+    /**
+     * @brief Gets the identifier for a user-assigned managed identity.
+     *
+     * @remark In the case of system-assigned managed identity, this will return an empty string.
+     */
+    std::string const& GetId() const { return m_id; }
+
+    /**
+     * @brief Gets the type of identifier used for the managed identity, depending on how it is
+     * configured.
+     */
+    ManagedIdentityIdType GetManagedIdentityIdType() const { return m_idType; }
+  };
+
+  /**
+   * @brief Options for managed identity credential.
+   *
+   */
+  struct ManagedIdentityCredentialOptions final : public Core::Credentials::TokenCredentialOptions
+  {
+    /**
+     * @brief Specifies the type of managed identity and its corresponding identifier, based on how
+     * it was configured.
+     */
+    ManagedIdentityType IdentityType;
+  };
+
+  /**
    * @brief Attempts authentication using a managed identity that has been assigned to the
    * deployment environment. This authentication type works in Azure VMs, App Service and Azure
    * Functions applications, as well as the Azure Cloud Shell. More information about configuring
@@ -70,6 +155,14 @@ namespace Azure { namespace Identity {
         std::string const& clientId = std::string(),
         Azure::Core::Credentials::TokenCredentialOptions const& options
         = Azure::Core::Credentials::TokenCredentialOptions());
+
+    /**
+     * @brief Constructs a Managed Identity Credential.
+     *
+     * @param options Options for token retrieval.
+     */
+    explicit ManagedIdentityCredential(
+        Azure::Identity::ManagedIdentityCredentialOptions const& options);
 
     /**
      * @brief Constructs an instance of ManagedIdentityCredential capable of authenticating a